requirements.txt
Paste==1.7.5.1
pyramid==1.5
pyramid-mako==1.0.2
waitress==0.8.9
gevent-socketio==0.3.5-rc2
